task_2/requirements.txt
attrs==21.2.0
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
PyJWT==1.7.1
pyrsistent==0.18.0
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
python-dotenv==0.18.0